pyproject.toml
# SPDX-FileCopyrightText: 2025 Andrey Kotlyar <guitar0.app@gmail.com>
#
# SPDX-License-Identifier: AGPL-3.0-or-later

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "guitar0-server"
version = "0.1.0"
description = "API Server for the mobile and the web clients"
authors = [
    {name = "Andrey Kotlyar", email = "kotlyar562@gmail.com"},
]
dependencies = [
    "Django>=5.2.8",
    "djangorestframework>=3.16.1",
    "pydantic-settings>=2.11.0",
    "dj-database-url>=3.0.1",
    "psycopg[binary]>=3.2.12",
    "pydantic>=2.12.4",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "AGPL-3.0-or-later"}

[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=6.1.0",
    "pre-commit>=4.3.0",
    "ruff>=0.14.3",
    "mypy>=1.18.2",
    "django-stubs>=5.2.7",
    "django-stubs-ext>=5.2.7",
    "djangorestframework-stubs>=3.16.5",
    "commitizen>=4.9.1",
    "reuse>=6.2.0",
]
test = [
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "factory-boy>=3.3.3",
    "model-bakery>=1.20.5",
    "faker>=37.12.0",
    "coverage[toml]>=7.11.0",
]


[tool.mypy]
python_version = "3.12"
plugins = [
    "pydantic.mypy",
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]
strict = true
ignore_missing_imports = false
warn_unused_configs = true
warn_return_any = true
warn_unreachable = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "apps.*.tests.factories"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"
warn_unused_configs = true
show_error_codes = true

[tool.ruff]
target-version = "py312"
line-length = 88
fix = true
extend-exclude = ["migrations"]
preview = true

[tool.ruff.lint]
extend-select = [
    "E", "W", "F", "I", "B", "N", "D",
    "C4", "UP", "PL", "DJ",
    "ANN", "SIM", "T10", "DTZ", "RUF"
]
ignore = [
    "D203", "D213", "D406", "D407", "D413", # docs black lines, etc.
    "RUF001", # every import on a new line
    "RUF002"  # cyrillic symbols, which have the same view, as latin
]
unfixable = ["F401"] # don't remove import without using

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["D100", "D101", "D102", "D103", "PLR2004"] # test files and functions without docs; numbers without constants
"**/models.py" = ["D106", "D105"] # docstring in nested classes (Meta), doc for __str__
"**/factories.py" = ["D106"]
"**/serializers.py" = ["D106"]
"__init__.py" = ["F401"] # imports without using in init files

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.pdm]
distribution = false

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = """
    --maxfail=1
    -ra
    -v
    --strict-markers
    --strict-config
    --cov
    --cov-fail-under=90
    --cov-report=term-missing
    --no-migrations
    -m 'not integration'
"""
testpaths = ["tests", "apps"]
markers = ["integration"]
filterwarnings = ["ignore:directory \"/run/secrets\" does not exist:UserWarning",]

[tool.coverage.run]
branch = true
source = ["."]
parallel = true
omit = ["*/migrations/*", "*/tests/*", "manage.py"]


[tool.coverage.report]
fail_under = 90
show_missing = true

[tool.commitizen]
name = "cz_conventional_commits"
version = "1.0.0"
tag_format = "v$version"
version_files = ["pyproject.toml"]

[tool.pdm.scripts]
setup = "pdm install --group dev --group test --group lint"
lint = "python scripts/lint.py"
license-check = "reuse lint"
license-add = """reuse annotate --recursive .
--skip-unrecognised
--copyright 'Andrey Kotlyar <guitar0.app@gmail.com>'
--license AGPL-3.0-or-later"""
